REFACTOR: move sound into component
